mqtt_listener/requirements.txt
paho-mqtt==1.5.1
pymongo==3.11.4
dnspython==2.1.0